support  mixed precision fp16 (WIP)

+++ src/transformers/models/segformer/modeling_tf_segformer.py
@@ -68,7 +68,7 @@ class TFSegformerDropPath(tf.keras.layers.Layer):
         if training:
             keep_prob = 1 - self.drop_path
             shape = (tf.shape(x)[0],) + (1,) * (len(tf.shape(x)) - 1)
-            random_tensor = keep_prob + tf.random.uniform(shape, 0, 1)
+            random_tensor = keep_prob + tf.random.uniform(shape, 0, 1, dtype=x.dtype)
             random_tensor = tf.floor(random_tensor)
             return (x / keep_prob) * random_tensor
         return x
